pyproject.toml
[project]
name = "stitching-images"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiocontextvars==0.2.2",
    "contextlib2==21.6.0",
    "exceptiongroup==1.3.0",
    "ipykernel==6.29.5",
    "ipython==8.12.3",
    "loguru>=0.7.3",
    "mock==5.2.0",
    "pillow==11.2.1",
]
